+++ hutool-core/src/test/java/cn/hutool/core/text/UnicodeUtilTest.java
@@ -25,4 +25,18 @@ public class UnicodeUtilTest {
 		String unicode = UnicodeUtil.toString(str);
 		Assert.assertEquals("aaaa&bbbb&cccc", unicode);
 	}
+	
+	@Test
+	public void convertTest3() {
+		String str = "aaa\\u111";
+		String res = UnicodeUtil.toString(str);
+		Assert.assertEquals("aaa\\u111", res);
+	}
+	
+	@Test
+	public void convertTest4() {
+		String str = "aaa\\U4e2d\\u6587\\u111\\urtyu\\u0026";
+		String res = UnicodeUtil.toString(str);
+		Assert.assertEquals("aaa中文\\u111\\urtyu&", res);
+	}
 }