+++ hutool-json/src/test/java/cn/hutool/json/JSONArrayTest.java
@@ -3,15 +3,13 @@ package cn.hutool.json;
 import java.util.ArrayList;
 
 import org.junit.Assert;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import cn.hutool.core.collection.CollUtil;
 import cn.hutool.core.io.FileUtil;
 import cn.hutool.core.lang.Console;
 import cn.hutool.core.util.CharsetUtil;
-import cn.hutool.json.JSONArray;
-import cn.hutool.json.JSONObject;
-import cn.hutool.json.JSONUtil;
 import cn.hutool.json.test.bean.Exam;
 
 /**
@@ -48,13 +46,10 @@ public class JSONArrayTest {
 		JSONObject obj0 = array.getJSONObject(0);
 		Exam exam = JSONUtil.toBean(obj0, Exam.class);
 		Assert.assertEquals("0", exam.getAnswerArray()[0].getSeq());
-		
-		for (JSONObject jsonObj : array.jsonIter()) {
-			Console.log(jsonObj.getInt("id"));
-		}
 	}
 	
 	@Test
+	@Ignore
 	public void parseBeanListTest() {
 		KeyBean b1 = new KeyBean();
 		b1.setAkey("aValue1");